+++ card/mod/bootstrap/set/all/bootstrap/form.rb
@@ -42,7 +42,7 @@ format :html do
   #     end
   #   end
   def bs_form opts={}, &block
-    BootstrapForm.render self, opts, &block
+    Card::Format::HtmlFormat::Bootstrap::Form.render self, opts, &block
   end
 
   # generate bootstrap form
@@ -54,69 +54,7 @@ format :html do
   #     end
   #   end
   def bs_horizontal_form *args, &block
-    BootstrapHorizontalForm.render self, *args, &block
+    Card::Format::HtmlFormat::Bootstrap::HorizontalForm.render self, *args, &block
   end
 
-  class BootstrapForm < BootstrapBuilder
-    def render_content *args
-      form *args, &@build_block
-    end
-
-    add_tag_method :form, nil, optional_classes: {
-                                 horizontal: "form-horizontal",
-                                 inline: "form-inline" }
-    add_div_method :group, "form-group"
-    add_tag_method :label, nil
-    add_tag_method :input, "form-control" do |opts, extra_args|
-      type, label = extra_args
-      prepend { label label, for: opts[:id] } if label
-      opts[:type] = type
-      opts
-    end
-
-    [:text, :password, :datetime, :"datetime-local", :date, :month, :time,
-     :week, :number, :email, :url, :search, :tel, :color].each do |tag|
-      add_tag_method tag, "form-control", attributes: { type: tag },
-                                          tag: :input do |opts, extra_args|
-        label, = extra_args
-        prepend { label label, for: opts[:id] } if label
-        opts
-      end
-    end
-  end
-
-  class BootstrapHorizontalForm < BootstrapForm
-    def left_col_width
-      @child_args.last && @child_args.last[0] || 2
-    end
-
-    def right_col_width
-      @child_args.last && @child_args.last[1] || 10
-    end
-
-    add_tag_method :form, "form-horizontal"
-
-    add_tag_method :label, "control-label" do |opts, extra_args|
-      prepend_class opts, "col-sm-#{left_col_width}"
-      opts
-    end
-
-    add_div_method :input, nil do |opts, extra_args|
-      type, label = extra_args
-      prepend { tag :label, nil, for: opts[:id] } if label
-      insert { inner_input opts.merge(type: type) }
-      { class: "col-sm-#{right_col_width}" }
-    end
-    add_tag_method :inner_input, "form-control", tag: :input
-    add_div_method :inner_checkbox, "checkbox"
-
-    add_div_method :checkbox, nil do |opts, extra_args|
-      inner_checkbox do
-        label do
-          inner_input "checkbox", extra_args.first, opts
-        end
-      end
-      { class: "col-sm-offset-#{left_col_width} col-sm-#{right_col_width}" }
-    end
-  end
 end
