+++ card/mod/bootstrap/set/all/bootstrap/helper.rb
@@ -52,80 +52,6 @@ format :html do
     '<li role="separator" class="divider"></li>'
   end
 
-  def breadcrumb items
-    content_tag :ol, class: "breadcrumb" do
-      items.map do |item|
-        content_tag :li, item, class: "breadcrumb-item"
-      end.join
-    end
-  end
-
-  # Options
-  # header: { content: String, brand: ( String | {name: , href: } ) }
-  def navbar id, opts={}
-    nav_opts = opts[:navbar_opts] || {}
-    nav_opts[:class] ||= opts[:class]
-    add_class nav_opts,
-              "navbar navbar-#{opts.delete(:navbar_type) || 'default'}"
-    header_opts = opts[:header] || {}
-    if opts[:toggle_align] == :left
-      opts[:toggle] = :hide
-      opts[:collapsed_content] ||= ""
-      opts[:collapsed_content] +=
-        navbar_toggle(
-          id, opts[:toggle], "pull-left navbar-link"
-        ).html_safe
-    end
-    wrap_with :nav, nav_opts do
-      [
-        navbar_header(id, header_opts.delete(:content),
-                      header_opts.reverse_merge(toggle: opts[:toggle])),
-        navbar_collapsed_content(opts[:collapsed_content]),
-        content_tag(:div, output(yield).html_safe,
-                    class: "collapse navbar-collapse",
-                    id: "navbar-collapse-#{id}")
-      ]
-    end
-  end
-
-  def navbar_collapsed_content content
-    content_tag(:div, content.html_safe, class: "container-fluid") if content
-  end
-
-  def navbar_header id, content="", opts={}
-    brand =
-      if opts[:brand]
-        if opts[:brand].is_a? String
-          "<a class='navbar-brand' href='#'>#{opts[:brand]}</a>"
-        else
-          link = opts[:brand][:href] || "#"
-          "<a class='navbar-brand' href='#{link}#'>#{opts[:brand][:name]}</a>"
-        end
-      end
-    wrap_with :div, class: "navbar-header" do
-      [
-        (navbar_toggle(id, opts[:toggle]) unless opts[:toggle] == :hide),
-        brand,
-        (content if content)
-      ]
-    end
-  end
-
-  def navbar_toggle id, content=nil, css_class=""
-    content ||= %(
-                  <span class="icon-bar"></span>
-                  <span class="icon-bar"></span>
-                  <span class="icon-bar"></span>
-                )
-    <<-HTML
-      <button type="button" class="navbar-toggle collapsed #{css_class}"
-              data-toggle="collapse" data-target="#navbar-collapse-#{id}">
-        <span class="sr-only">Toggle navigation</span>
-        #{content}
-      </button>
-    HTML
-  end
-
   def split_button button, args={}
     items = yield
     args[:situation] ||= "primary"
@@ -167,69 +93,6 @@ format :html do
         i_opts ||= default_item_options
         content_tag :li, i_content.html_safe, i_opts
       end.join "\n"
-    end
-  end
-
-  def accordion_group list, collapse_id=card.cardname.safe_key, args={}
-    accordions = ""
-    index = 1
-    case list
-    when Array then accordions = list.join
-    when String then accordions = list
-    else
-      list.each_pair do |title, content|
-        accordions << accordion(title, content, "#{collapse_id}-#{index}")
-        index += 1
-      end
-    end
-    add_class args, "panel-group"
-
-    content_tag :div, accordions.html_safe,
-                class: args[:class],
-                id: "accordion-#{collapse_id}",
-                role: "tablist",
-                "aria-multiselectable" => "true"
-  end
-
-  def accordion title, content, collapse_id=card.cardname.safe_key
-    accordion_content =
-      case content
-      when Hash  then accordion_group(content, collapse_id)
-      when Array then content.present? && list_group(content)
-      when String then content
-      end
-    <<-HTML.html_safe
-      <div class="panel panel-default">
-        #{accordion_panel(title, accordion_content, collapse_id)}
-      </div>
-    HTML
-  end
-
-  def accordion_panel title, body, collapse_id, panel_heading_link=false
-    if body
-      <<-HTML
-        <div class="panel-heading" role="tab" id="heading-#{collapse_id}">
-          <h4 class="panel-title">
-            <a data-toggle="collapse" data-parent="#accordion-#{collapse_id}" \
-               href="##{collapse_id}" aria-expanded="true" \
-               aria-controls="#{collapse_id}">
-              #{title}
-            </a>
-          </h4>
-        </div>
-        <div id="#{collapse_id}" class="panel-collapse collapse" \
-               role="tabpanel" aria-labelledby="heading-#{collapse_id}">
-          <div class="panel-body">
-            #{body}
-          </div>
-        </div>
-      HTML
-    else
-      <<-HTML
-        <li class="list-group-item">
-          <h4 class="panel-title">#{title}</h4>
-        </li>
-      HTML
     end
   end
 end
