+++ card/mod/bootstrap/set/all/bootstrap/panel.rb
@@ -1,11 +1,5 @@
 format :html do
   def bs_panel opts={}, &block
-    BootstrapPanel.render self, opts, &block
-  end
-
-  class BootstrapPanel < BootstrapBuilder
-    add_div_method :panel, "panel panel-default"
-    add_div_method :heading, "panel-heading"
-    add_div_method :body, "panel-body"
+    Bootstrap::Panel.render self, opts, &block
   end
 end
