+++ card/mod/history/set/all/history.rb
@@ -85,11 +85,12 @@ event :rollback_actions, :prepare_to_validate,
     end
   end
   Env.params["action_ids"] = nil
+  binding.pry
   update_attributes! revision
-  rollback_actions.each do |action|
-    # rollback file and image cards
-    action.card.try :rollback_to, action
-  end
+  # rollback_actions.each do |action|
+  #   # rollback file and image cards
+  #   action.card.try :rollback_to, action
+  # end
   clear_drafts
   abort :success
 end
@@ -140,7 +141,7 @@ end
 format :html do
   view :history do |args|
     frame args.merge(body_class: "history-slot list-group", content: true) do
-      [_render_history_legend, _render_history_acts(args)]
+      [history_legend, history_acts(args)]
     end
   end
 
@@ -148,22 +149,22 @@ format :html do
     args[:optional_toolbar] ||= :show
   end
 
-  view :history_acts do |args|
+  def history_acts args
     page = params["page"] || 1
     args[:acts] = card.intrusive_acts.page(page).per(ACTS_PER_PAGE)
     _render_act_list args
   end
 
-  view :history_legend do
+  def history_legend with_drafts=true
     intr = card.intrusive_acts.page(params["page"]).per(ACTS_PER_PAGE)
-    legend = render_haml intr: intr do
+    legend = render_haml intr: intr, with_drafts: with_drafts do
       <<-HAML.strip_heredoc
         .history-header
           %span.slotter
             = paginate intr, remote: true, theme: 'twitter-bootstrap-3'
           %div.history-legend
             %span.pull-left
-              = action_legend
+              = action_legend with_drafts
             %span.pull-right
               Content changes:
               %span
@@ -181,182 +182,24 @@ format :html do
     end
   end
 
-  def action_legend
-    legend = [:create, :update, :delete].map do |action_type|
+  def action_legend with_drafts
+    types = [:create, :update, :delete]
+    legend = types.map do |action_type|
                "#{action_icon(action_type)} #{action_type}d"
-              end.join " | "
+    end.join " | "
+    if with_drafts
+      legend << " | #{action_icon(:draft)} unsaved draft"
+    end
     "Actions: #{legend}"
   end
 
-  def action_list args
-    act = args[:act]
-    actions =
-      if act_context(args) == :absolute
-        act.actions
-      else
-        act.actions_affecting(card)
-      end
-    actions.select { |a| a.card && a.card.ok?(:read) }
-    # FIXME: should not need to test for presence of card here.
-  end
-
-  def hide_diff?
-    params["hide_diff"].to_s.strip == "true"
-  end
-
-  def action_view
-    (params["action_view"] || "summary").to_sym
-  end
-
-  view :act_header do |_args|
-    %(<h5 class="act-header">#{link_to_card card}</h5>)
-  end
-
-  view :act_metadata do |args|
-    render_haml args.merge(card: card, args: args) do
-      <<-HAML.strip_heredoc
-        - unless act_context == :absolute
-          .nr
-            = '#' + act_seq.to_s
-        .title
-          .actor
-            = link_to_card act.actor
-          .time.timeago
-            = time_ago_in_words(act.acted_at)
-            ago
-            - if act.id == card.last_act.id
-              %em.label.label-info Current
-            - if action_view == :expanded
-              - unless act.id == card.last_act.id
-                = rollback_link act.actions_affecting(card)
-              = show_or_hide_changes_link args
-      HAML
-    end
-  end
-
-  view :action_summary do |args|
-    view_action :summary, args
-  end
-
-  view :action_expanded do |args|
-    view_action :expanded, args
-  end
-
-  def view_action action_view, args
-    action = args[:action] || card.last_action
-    hide_diff = args[:hide_diff] || hide_diff?
-    return trashed_view(action) if action.action_type == :delete
-    render_haml action: action,
-                action_view: action_view,
-                name_diff: name_diff(action, hide_diff),
-                type_diff: type_diff(action, hide_diff),
-                content_diff: content_diff(action, action_view, hide_diff) do
-      <<-HAML.strip_heredoc
-        .action
-          .summary
-            %span.ampel
-              = glyphicon 'minus-sign', (action.red? ? 'deleted-mark' : 'diff-invisible')
-              = glyphicon 'plus-sign', (action.green? ? 'added-mark' : 'diff-invisible')
-            = wrap_diff :name, name_diff
-            = wrap_diff :type, type_diff
-            -if content_diff && action_view == :summary
-              = glyphicon 'arrow-right', 'arrow'
-              = wrap_diff :content, content_diff
-          -if content_diff and action_view == :expanded
-            .expanded
-              = wrap_diff :content, content_diff
-      HAML
-    end
-  end
-
-  def trashed_view action
-    render_haml action: action do
-      <<-HAML.strip_heredoc
-        .action
-          .summary
-            %span.ampel
-              = glyphicon 'trash', 'deleted-mark'
-            = wrap_diff :name, action.card.name, ('label label-default' if action.card != card)
-      HAML
-    end
-  end
-
-  def name_diff action, hide_diff
-    working_name = name_changes action, hide_diff
-    if action.card == card
-      working_name
-    else
-      link_to_view(
-        :related, working_name,
-        path: { related: { view: "history", name: action.card.name } },
-        remote: true,
-        class: "slotter label label-default",
-        "data-slot-selector" => ".card-slot.history-view"
-      )
-    end
-  end
-
-  def type_diff action, hide_diff
-    action.new_type? && type_changes(action, hide_diff)
-  end
-
-  def content_diff action, action_view, hide_diff
-    diff = action.new_content? &&
-           action.card.format.render_content_changes(
-             action: action, diff_type: action_view, hide_diff: hide_diff
-           )
-    return "<i>empty</i>" unless diff.present?
-    diff
-  end
-
-  def wrap_diff field, content, extra_class=nil
-    return "" unless content.present?
-    %(
-       <span class="#{field}-diff #{extra_class}">
-       #{content}
-       </span>
-    )
-  end
-
-  def name_changes action, hide_diff=false
-    old_name = (name = action.previous_value :name) && showname(name).to_s
-    if action.new_name?
-      new_name = showname(action.value(:name)).to_s
-      if hide_diff
-        new_name
-      else
-        Card::Content::Diff.complete(old_name, new_name)
-      end
-    else
-      old_name
-    end
-  end
-
-  def type_changes action, hide_diff=false
-    change = hide_diff ? action.value(:cardtype) : action.cardtype_diff
-    "(#{change})"
-  end
-
   view :content_changes do |args|
+    action = args[:action]
     if args[:hide_diff]
-      args[:action].raw_view
+      action.raw_view
     else
-      args[:action].content_diff(args[:diff_type])
+      action.content_diff(args[:diff_type])
     end
-  end
-
-  def fold_or_unfold_link args
-    act_id = args[:act].id
-    action_view = args[:action_view] == :expanded ? :summary : :expanded
-    arrow_dir = args[:action_view] == :expanded ? "arrow-down" : "arrow-right"
-
-    link_to_view :act, "", class: "slotter revision-#{act_id} #{arrow_dir}",
-                           path: { act_id:      act_id,
-                                   act_seq:     args[:act_seq],
-                                   hide_diff:   args[:hide_diff],
-                                   act_context: args[:act_context],
-                                   action_view: action_view,
-                                   look_in_trash: true }
   end
 end
