+++ card/mod/machines/set/right/machine_input.rb
@@ -1,3 +1,7 @@
 def followable?
   false
 end
+
+def history?
+  false
+end
